+++ pandapower/opf/make_objective.py
@@ -7,9 +7,8 @@
 # Copyright (c) 2016-2018 by University of Kassel and Fraunhofer Institute for Energy Economics
 # and Energy System Technology (IEE), Kassel. All rights reserved.
 
-
+from numpy import zeros, array, concatenate, power, ndarray
 import pandas as pd
-from numpy import zeros, array, concatenate, power
 from pandapower.idx_cost import MODEL, NCOST, COST
 
 
@@ -29,7 +28,7 @@ def _make_objective(ppci, net):
     """
     # Determine duplicated cost data
     all_costs = net.polynomial_cost[['type', 'element', 'element_type']].append(
-            net.piecewise_linear_cost[['type', 'element', 'element_type']])
+        net.piecewise_linear_cost[['type', 'element', 'element_type']])
     duplicates = all_costs.loc[all_costs.duplicated()]
     if duplicates.shape[0]:
         raise ValueError("There are elements with multipy costs.\nelement_types: %s\n"
@@ -81,8 +80,8 @@ def _make_objective(ppci, net):
         if len(net.piecewise_linear_cost):
             for cost_type in ["p", "q"]:
                 if (net.piecewise_linear_cost.type == cost_type).any():
-                    costs = net.piecewise_linear_cost[net.piecewise_linear_cost.type == cost_type].reset_index(
-                        drop=True)
+                    costs = net.piecewise_linear_cost[net.piecewise_linear_cost.type ==
+                                                      cost_type].reset_index(drop=True)
 
                     if cost_type == "q":
                         shift_idx = ng
@@ -115,22 +114,22 @@ def _make_objective(ppci, net):
                                 el_is = net[el].loc[net[el].in_service & net[el].index.isin(
                                     costs.loc[costs.element_type == el].element)].index
                             else:
-                                el_is = net[el].loc[net[el].controllable & net[el].in_service & net[el].index.isin(
-                                    costs.loc[costs.element_type == el].element)].index
+                                el_is = net[el].loc[net[el].controllable & net[el].in_service &
+                                                    net[el].index.isin(costs.loc[costs.element_type
+                                                                                 == el].element)].index
 
                             p = costs.loc[(costs.element_type == el) & (
-                                    costs.element.isin(el_is))].p.reset_index(drop=True)
+                                costs.element.isin(el_is))].p.reset_index(drop=True)
                             f = costs.loc[(costs.element_type == el) & (
-                                    costs.element.isin(el_is))].f.reset_index(drop=True)
+                                costs.element.isin(el_is))].f.reset_index(drop=True)
 
                             if len(p) > 0:
                                 p = concatenate(p)
                                 f = concatenate(f)
                                 # gencost indices
                                 elements = idx[el_is] + shift_idx
-
-                                ppci["gencost"][elements, COST:COST+n_piece_lin_coefficients:2] = p
-
+                                ppci["gencost"][elements, COST:COST +
+                                                n_piece_lin_coefficients:2] = p
                                 # gencost for storages: positive costs in pandapower per definition
                                 # --> storage gencosts are similar to sgen gencosts
                                 if el in ["load", "dcline"]:
@@ -138,15 +137,17 @@ def _make_objective(ppci, net):
                                                     n_piece_lin_coefficients+1:2] = - f * 1e3
                                 else:
                                     ppci["gencost"][elements, COST+1:COST+n_piece_lin_coefficients +
-                                                    1:2] = f * 1e3 * sign_corr
+                                                    1:2] = - f * 1e3 * sign_corr
 
-                                ppci["gencost"][elements, NCOST] = n_coefficients / 2
+                                ppci["gencost"][elements,
+                                                NCOST] = n_coefficients / 2
                                 ppci["gencost"][elements, MODEL] = 1
 
         if len(net.polynomial_cost):
             for cost_type in ["p", "q"]:
                 if (net.polynomial_cost.type == cost_type).any():
-                    costs = net.polynomial_cost[net.polynomial_cost.type == cost_type].reset_index(drop=True)
+                    costs = net.polynomial_cost[net.polynomial_cost.type == cost_type].reset_index(
+                        drop=True)
 
                     if cost_type == "q":
                         shift_idx = ng
@@ -184,7 +185,7 @@ def _make_objective(ppci, net):
                             c = costs.loc[(costs.element_type == el) &
                                           (costs.element.isin(el_is))].c.reset_index(drop=True)
 
-                            if len(c) > 0:
+                            if len(c) > 0 and isinstance(idx, ndarray):
                                 c = concatenate(c)
                                 n_c = c.shape[1]
                                 c = c * power(1e3, array(range(n_c))[::-1])
@@ -198,11 +199,14 @@ def _make_objective(ppci, net):
                                 # gencost for storages: positive costs in pandapower per definition
                                 # --> storage gencosts are similar to sgen gencosts
                                 if el in ["load", "dcline"]:
-                                    ppci["gencost"][elements, COST:(COST + n_c):] = - c
+                                    ppci["gencost"][elements,
+                                                    COST:(COST + n_c):] = - c
                                 else:
-                                    ppci["gencost"][elements, -n_c:n_gencost] = c * sign_corr
+                                    ppci["gencost"][elements, -
+                                                    n_c:n_gencost] = - c * sign_corr
 
-                                ppci["gencost"][elements, NCOST] = n_coefficients
+                                ppci["gencost"][elements,
+                                                NCOST] = n_coefficients
                                 ppci["gencost"][elements, MODEL] = 2
 
     else: